Refactor quiz.js
